# api/pkg/dbapi/types.go
// Package dbapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package dbapi

import (
	"encoding/json"

	"github.com/oapi-codegen/runtime"
)

// ApiResponse defines model for ApiResponse.
type ApiResponse struct {
	Message string `json:"message"`
}

// GenreResponse defines model for GenreResponse.
type GenreResponse struct {
	GenreId int    `json:"genre_id"`
	Name    string `json:"name"`
}

// Group defines model for Group.
type Group struct {
	GroupId   int    `json:"group_id"`
	GroupName string `json:"group_name"`
}

// GroupResponse defines model for GroupResponse.
type GroupResponse struct {
	GroupId int    `json:"group_id"`
	Name    string `json:"name"`
}

// HTTPValidationError defines model for HTTPValidationError.
type HTTPValidationError struct {
	Detail *[]ValidationError `json:"detail,omitempty"`
}

// MovieReccomendationResponse defines model for MovieReccomendationResponse.
type MovieReccomendationResponse struct {
	Accuracy float32 `json:"accuracy"`
	MovieId  string  `json:"movie_id"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	Mail   string `json:"mail"`
	Name   string `json:"name"`
	UserId int    `json:"user_id"`
}

// ValidationError defines model for ValidationError.
type ValidationError struct {
	Loc  []ValidationError_Loc_Item `json:"loc"`
	Msg  string                     `json:"msg"`
	Type string                     `json:"type"`
}

// ValidationErrorLoc0 defines model for .
type ValidationErrorLoc0 = string

// ValidationErrorLoc1 defines model for .
type ValidationErrorLoc1 = int

// ValidationError_Loc_Item defines model for ValidationError.loc.Item.
type ValidationError_Loc_Item struct {
	union json.RawMessage
}

// CreateGroupGroupsPostParams defines parameters for CreateGroupGroupsPost.
type CreateGroupGroupsPostParams struct {
	GpName string `form:"gp_name" json:"gp_name"`
}

// AddGroupRecommendationsGroupsGroupIdRecommendationsPostParams defines parameters for AddGroupRecommendationsGroupsGroupIdRecommendationsPost.
type AddGroupRecommendationsGroupsGroupIdRecommendationsPostParams struct {
	MovieId  int     `form:"movie_id" json:"movie_id"`
	Accuracy float32 `form:"accuracy" json:"accuracy"`
}

// CreateUserUsersPostParams defines parameters for CreateUserUsersPost.
type CreateUserUsersPostParams struct {
	Mail string `form:"mail" json:"mail"`
	Name string `form:"name" json:"name"`
}

// AddUserInfosUsersUserIdPostParams defines parameters for AddUserInfosUsersUserIdPost.
type AddUserInfosUsersUserIdPostParams struct {
	Name string `form:"name" json:"name"`
}

// AddMovieToWatchedUsersUserIdMoviesMovieIdPostParams defines parameters for AddMovieToWatchedUsersUserIdMoviesMovieIdPost.
type AddMovieToWatchedUsersUserIdMoviesMovieIdPostParams struct {
	Rating int `form:"rating" json:"rating"`
}

// AddRecommendationsUsersUserIdRecommendationsPostParams defines parameters for AddRecommendationsUsersUserIdRecommendationsPost.
type AddRecommendationsUsersUserIdRecommendationsPostParams struct {
	MovieId  int     `form:"movie_id" json:"movie_id"`
	Accuracy float32 `form:"accuracy" json:"accuracy"`
}

// AsValidationErrorLoc0 returns the union data inside the ValidationError_Loc_Item as a ValidationErrorLoc0
func (t ValidationError_Loc_Item) AsValidationErrorLoc0() (ValidationErrorLoc0, error) {
	var body ValidationErrorLoc0
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromValidationErrorLoc0 overwrites any union data inside the ValidationError_Loc_Item as the provided ValidationErrorLoc0
func (t *ValidationError_Loc_Item) FromValidationErrorLoc0(v ValidationErrorLoc0) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeValidationErrorLoc0 performs a merge with any union data inside the ValidationError_Loc_Item, using the provided ValidationErrorLoc0
func (t *ValidationError_Loc_Item) MergeValidationErrorLoc0(v ValidationErrorLoc0) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsValidationErrorLoc1 returns the union data inside the ValidationError_Loc_Item as a ValidationErrorLoc1
func (t ValidationError_Loc_Item) AsValidationErrorLoc1() (ValidationErrorLoc1, error) {
	var body ValidationErrorLoc1
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromValidationErrorLoc1 overwrites any union data inside the ValidationError_Loc_Item as the provided ValidationErrorLoc1
func (t *ValidationError_Loc_Item) FromValidationErrorLoc1(v ValidationErrorLoc1) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeValidationErrorLoc1 performs a merge with any union data inside the ValidationError_Loc_Item, using the provided ValidationErrorLoc1
func (t *ValidationError_Loc_Item) MergeValidationErrorLoc1(v ValidationErrorLoc1) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t ValidationError_Loc_Item) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *ValidationError_Loc_Item) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}