# db/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/{user_id}/movies":{"get":{"tags":["Movies of Users"],"summary":"Get Seen Movies","operationId":"get_seen_movies_users__user_id__movies_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"title":"Response Get Seen Movies Users  User Id  Movies Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/movies/{movie_id}":{"post":{"tags":["Movies of Users"],"summary":"Add Movie To Watched","operationId":"add_movie_to_watched_users__user_id__movies__movie_id__post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"movie_id","in":"path","required":true,"schema":{"type":"integer","title":"Movie Id"}},{"name":"rating","in":"query","required":true,"schema":{"type":"integer","title":"Rating"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Movies of Users"],"summary":"Remove Movie From Watched","operationId":"remove_movie_from_watched_users__user_id__movies__movie_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"movie_id","in":"path","required":true,"schema":{"type":"integer","title":"Movie Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/recommendations":{"get":{"tags":["Recommendations"],"summary":"Get Recommendations","operationId":"get_recommendations_users__user_id__recommendations_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieReccomendationResponse"},"title":"Response Get Recommendations Users  User Id  Recommendations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Recommendations"],"summary":"Add Recommendations","operationId":"add_recommendations_users__user_id__recommendations_post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"movie_id","in":"query","required":true,"schema":{"type":"integer","title":"Movie Id"}},{"name":"accuracy","in":"query","required":true,"schema":{"type":"number","title":"Accuracy"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/tastes":{"get":{"tags":["Tastes of Users"],"summary":"Get Tastes","operationId":"get_tastes_users__user_id__tastes_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GenreResponse"},"title":"Response Get Tastes Users  User Id  Tastes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/tastes/{genre_id}":{"post":{"tags":["Tastes of Users"],"summary":"Add Taste For User","operationId":"add_taste_for_user_users__user_id__tastes__genre_id__post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Tastes of Users"],"summary":"Remove Taste From User","operationId":"remove_taste_from_user_users__user_id__tastes__genre_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{email}/bymail":{"get":{"tags":["Users"],"summary":"Get User Infos By Mail","operationId":"get_user_infos_by_mail_users__email__bymail_get","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"tags":["Users"],"summary":"Get User Infos","operationId":"get_user_infos_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Users"],"summary":"Add User Infos","operationId":"add_user_infos_users__user_id__post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Remove User","operationId":"remove_user_users__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"post":{"tags":["Users"],"summary":"Create User","operationId":"create_user_users_post","parameters":[{"name":"mail","in":"query","required":true,"schema":{"type":"string","title":"Mail"}},{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/{user_id}/groups":{"get":{"tags":["Users"],"summary":"Get User Groups","operationId":"get_user_groups_user__user_id__groups_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Group"},"title":"Response Get User Groups User  User Id  Groups Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups":{"post":{"tags":["Groups"],"summary":"Create Group","operationId":"create_group_groups_post","parameters":[{"name":"gp_name","in":"query","required":true,"schema":{"type":"string","title":"Gp Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}":{"get":{"tags":["Groups"],"summary":"Get Group Infos","operationId":"get_group_infos_groups__group_id__get","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}/users":{"get":{"tags":["Users from Groups"],"summary":"Get Group Users","operationId":"get_group_users_groups__group_id__users_get","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"integer"},"title":"Response Get Group Users Groups  Group Id  Users Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}/users/{user_id}":{"post":{"tags":["Users from Groups"],"summary":"Add User To Group","operationId":"add_user_to_group_groups__group_id__users__user_id__post","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}},{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users from Groups"],"summary":"Remove User From Group","operationId":"remove_user_from_group_groups__group_id__users__user_id__delete","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}},{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}/recommendations":{"get":{"tags":["Recommendations"],"summary":"Get Group Recommendations","operationId":"get_group_recommendations_groups__group_id__recommendations_get","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieReccomendationResponse"},"title":"Response Get Group Recommendations Groups  Group Id  Recommendations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Recommendations"],"summary":"Add Group Recommendations","operationId":"add_group_recommendations_groups__group_id__recommendations_post","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}},{"name":"movie_id","in":"query","required":true,"schema":{"type":"integer","title":"Movie Id"}},{"name":"accuracy","in":"query","required":true,"schema":{"type":"number","title":"Accuracy"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}/tastes":{"get":{"tags":["Group tastes"],"summary":"Get Group Tastes","operationId":"get_group_tastes_groups__group_id__tastes_get","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GenreResponse"},"title":"Response Get Group Tastes Groups  Group Id  Tastes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}/tastes/{genre_id}":{"post":{"tags":["Group tastes"],"summary":"Add Taste For Group","operationId":"add_taste_for_group_groups__group_id__tastes__genre_id__post","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}},{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Group tastes"],"summary":"Remove Taste From Group","operationId":"remove_taste_from_group_groups__group_id__tastes__genre_id__delete","parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}},{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ApiResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"ApiResponse"},"GenreResponse":{"properties":{"genre_id":{"type":"integer","title":"Genre Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["genre_id","name"],"title":"GenreResponse"},"Group":{"properties":{"group_id":{"type":"integer","title":"Group Id"},"group_name":{"type":"string","title":"Group Name"}},"type":"object","required":["group_id","group_name"],"title":"Group"},"GroupResponse":{"properties":{"group_id":{"type":"integer","title":"Group Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["group_id","name"],"title":"GroupResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MovieReccomendationResponse":{"properties":{"movie_id":{"type":"string","title":"Movie Id"},"accuracy":{"type":"number","title":"Accuracy"}},"type":"object","required":["movie_id","accuracy"],"title":"MovieReccomendationResponse"},"UserResponse":{"properties":{"user_id":{"type":"integer","title":"User Id"},"name":{"type":"string","title":"Name"},"mail":{"type":"string","title":"Mail"}},"type":"object","required":["user_id","name","mail"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"Movies of Users","description":"Manage the movies seen by the user."},{"name":"Recommendations","description":"Manage the recommendations for the users and the groups."},{"name":"Tastes of Users","description":"Manage the tastes in genre of a specific user."},{"name":"Users","description":"CRUD for users."},{"name":"Groups","description":"CR for groups."},{"name":"Users from Groups","description":"CRUD for users in a group."},{"name":"Group tastes","description":"Get the tastes for a group."}]}